requirements.txt
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
pillow==11.1.0
django-filter==25.1
drf-yasg==1.21.9
psycopg2-binary==2.9.10
firebase-admin==6.7.0
django-environ==0.12.0
django-cleanup==9.0.0
django-flags==5.0.13
sentry-sdk[django]==2.27.0
gunicorn==23.0.0
django-storages==1.14.5
boto3==1.38.26
mysqlclient==2.2.7
openai==1.93.0
